requirements.txt
numpy
tensorflow
deeppavlov
nltk
spacy
termcolor
pandas
ufal.udpipe
stopwordsiso
polyglot
pyicu
pycld2
morfessor
psycopg2-binary
pyAesCrypt
paramiko
cryptography==2.5
boto3

